examples/python-langchain/pyproject.toml
[project]
name = "sigil-langchain-weather-example"
version = "0.1.0"
description = "Minimal FastAPI + LangChain weather agent instrumented with Sigil and OpenTelemetry."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-dotenv>=1.0",
    "pydantic>=2.8",

    "langchain>=1.2.15",
    "langchain-core>=1.3.0",
    "langchain-anthropic>=1.4.1",

    "anthropic>=0.96",

    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48b0",

    "sigil-sdk>=0.1.2",
    "sigil-sdk-langchain>=0.1.2",
]

[tool.uv.sources]
# Point at the local packages in this monorepo so the example runs against
# the code in this checkout. Remove these two entries to consume the
# published PyPI releases instead.
sigil-sdk = { path = "../../python", editable = true }
sigil-sdk-langchain = { path = "../../python-frameworks/langchain", editable = true }

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pyright]
venvPath = "."
venv = ".venv"
extraPaths = ["../../python", "../../python-frameworks/langchain"]
